[MMS] Use chans instead of normal addresses for auto-config
Using the Bitmessage mechanism of chans instead of normal
addresses as the basis for the auto-config mechanism of
the MMS has advantages:

They are faster because no initial key exchange is needed -
everybody who knows the chan name knows already the key
(or better said, can derive the key from the chan name).

Especially in connection with the PyBitmessage app that the MMS
uses chans have another advantage over normal addresses: It's
possible to delete them manually if anything goes wrong and the MMS
has no chance to delete them automatically.

Just for completeness' sake it should be noted that one should not
bring this change into service exactly in the middle of an
already-running auto-config process.

Ref: https://github.com/monero-project/monero/pull/5778/commits/8703aa59db18918d6d6687c25a5461be3521adcf