[ZMQ] Dropping cppzmq dependency, adding some zmq utils
This was adapted from the light wallet server patch after an issue
was posted against cppzmq (its not explicitly listed as a dependency).
This dependency on the cppzmq header is dropped, and instead
`unique_ptr` and a few smaller functions are used. cppzmq requires
the usage of exceptions for errors (and does something funky with
EAGAIN), whereas this lets the caller decide between error codes
and exceptions.

Also a few changes:

* ZMQ server now has a max input message size of 10 MiB
* ZMQ server handles a message sent in parts
* ZMQ server no longer polls for shutdown every second, `zmq_term`
is used to signal shutdown in blocking read/write calls.
* ZMQ server now automatically retries on `EINTR` errors during
read/write calls  (see above point).

This was verified with some unit tests and against
the light-wallet-server.

Ref: https://github.com/monero-project/monero/pull/5818