[P2P] syncing delay at the end of chain update
Change the default block_sync_size from 50 to 1 starting from 90% of the chain downloaded instead of 95% as per my previous commit. The delay at the end has a correlation with the distance between the end chain height and the latest checkpoint. So till we update checkpoints the furthest the chain top goes from the checkpoint the earliest the delay starts. (when i committed a month ago delay started at 96% of the chain downloaded now it starts at 94%)